Revert unnecessary changes
put_nth_field(): Assume that the field is materialized in the record.
Key columns can never be instantly added, and the only caller
fill_lock_data() is restricting the n_fields to the key columns.

fts_fetch_doc_from_rec(): Instant ADD COLUMN is never invoked on
the internal tables that are created for implementing FULLTEXT INDEX
for InnoDB tables. Use the low-level physical access to the fields.